+++ frontend/src/AssumePartyPage.tsx
@@ -22,13 +22,13 @@ import {
   CardContent,
   CardHeader,
 } from "@mui/material";
-import PeopleAltIcon from "@mui/icons-material/PeopleAlt";
+import PersonIcon from "@mui/icons-material/Person";
 import { useState, useEffect } from "react";
 import { docsURL } from "./httpConfig";
 import { ScopesField } from "./components/scopes";
+import { useTheme } from "@mui/material/styles";
 
 export const AssumePartyPage = () => {
-  const [loading, setLoading] = useState(false);
   const [unAssumed, setUnAssumed] = useState(false);
   const notify = useNotify();
   const redirect = useRedirect();
@@ -59,6 +59,8 @@ export const AssumePartyPage = () => {
 
   const AssumePartyButton = ({ field }: any) => {
     const record = useRecordContext()!;
+    const elhubTheme = useTheme();
+    const [loading, setLoading] = useState(false);
     const assumeParty = async () => {
       setLoading(true);
       return login({ party_id: record[field] })
@@ -84,11 +86,21 @@ export const AssumePartyPage = () => {
     };
 
     return (
-      <Button label="Act on behalf of this party" onClick={assumeParty}>
-        <>
-          <PeopleAltIcon />
-          {loading && <CircularProgress size={25} thickness={2} />}
-        </>
+      <Button
+        color="secondary"
+        size="medium"
+        onClick={assumeParty}
+        sx={{
+          fontWeight: "bold",
+          border: "2px solid transparent",
+          ":hover": {
+            border: `2px solid ${elhubTheme.palette.primary.main}`,
+          },
+        }}
+      >
+        {loading && <CircularProgress size={25} thickness={2} sx={{ mr: 1 }} />}
+        <PersonIcon sx={{ mr: 1 }} />
+        Act on behalf of {record.name}
       </Button>
     );
   };
@@ -122,7 +134,7 @@ export const AssumePartyPage = () => {
           actions={false}
           perPage={25}
           sort={{ field: "id", order: "ASC" }}
-          filter={{ entity_id: identity.data!.entityID }}
+          filter={{ entity_id: identity.data?.entityID ?? "" }}
           empty={
             <Box textAlign="center" m={1}>
               <Typography variant="h5">
@@ -134,7 +146,6 @@ export const AssumePartyPage = () => {
         >
           <Datagrid bulkActionButtons={false}>
             <TextField label="ID" source="party_id" />
-            <AssumePartyButton field="party_id" />
             <ReferenceField
               label="Name"
               source="party_id"
@@ -145,6 +156,7 @@ export const AssumePartyPage = () => {
             </ReferenceField>
             <ReferenceField
               label="Type"
+              link={false}
               source="party_id"
               reference="party"
               sortable={false}
@@ -152,6 +164,15 @@ export const AssumePartyPage = () => {
               <TextField source="type" />
             </ReferenceField>
             <ScopesField source="scopes" />
+            <ReferenceField
+              label="Assume party"
+              source="party_id"
+              reference="party"
+              link={false}
+              sortable={false}
+            >
+              <AssumePartyButton field="id" />
+            </ReferenceField>
           </Datagrid>
         </List>
       )}
@@ -163,7 +184,7 @@ export const AssumePartyPage = () => {
       ) : (
         <ResourceContextProvider value="party">
           <List
-            title={false}
+            title=""
             actions={false}
             perPage={5}
             sort={{ field: "id", order: "ASC" }}
@@ -175,11 +196,19 @@ export const AssumePartyPage = () => {
             }
             disableSyncWithLocation
           >
-            <Datagrid bulkActionButtons={false}>
+            <Datagrid bulkActionButtons={false} rowClick={false}>
               <TextField label="ID" source="id" />
-              <AssumePartyButton field="id" />
               <TextField source="name" />
               <TextField source="type" />
+              <ReferenceField
+                label="Assume party"
+                source="id"
+                reference="party"
+                link={false}
+                sortable={false}
+              >
+                <AssumePartyButton field="id" />
+              </ReferenceField>
             </Datagrid>
           </List>
         </ResourceContextProvider>